ISPN-14545 SIFS Compactor does not properly shut down but the index thinks it is okay

+++ core/src/main/java/org/infinispan/persistence/sifs/FileProvider.java
@@ -296,8 +296,11 @@ public class FileProvider {
          }
          if (!recordQueue.isEmpty()) throw new IllegalStateException();
          if (!openFiles.isEmpty()) throw new IllegalStateException();
-         for (File file : dataDir.listFiles()) {
-            Files.delete(file.toPath());
+         File[] files = dataDir.listFiles();
+         if (files != null) {
+            for (File file : files) {
+               Files.delete(file.toPath());
+            }
          }
       } finally {
          lock.writeLock().unlock();
